refactor(services/shop): reacomodo del código de register y login en los controladores de shop. crea
