+++ services/shop/src/v1/controllers/user/register.js
@@ -1,4 +1,5 @@
-import { hashPassword, userModel } from '../../../../../../utils/middlewares/auth/bcrypt.js'
+import bcrypt from 'bcrypt'
+import { hashPassword } from '../../../../../../utils/middlewares/auth/authPassport.js'
 import { User } from '../../database/connection/connectionDB.js'
 
 // se debe utilizar el modelo User de cada servicio
@@ -10,10 +11,10 @@ const register = async (req, res, next) => {
     if (!password || !email) return res.json({ status: 400, message: 'ingrese los campos obligatorios' })
 
     else if (email && password && !name) {
-      const [user, isCreated] = await User.findOrCreate({ where: { email }, defaults: { email, password: hashPassword(password) } })
+      const [user, isCreated] = await User.findOrCreate({ where: { email }, defaults: { email, password: hashPassword(password, bcrypt) } })
       isCreated ? res.json({ status: 200, message: 'usuario registrado correctamente', user }) : res.json({ status: 404, message: 'esta cuenta ya existe' })
     } else {
-      const [user, isCreated] = await User.findOrCreate({ where: { email }, defaults: { name, email, password: hashPassword(password) } })
+      const [user, isCreated] = await User.findOrCreate({ where: { email }, defaults: { name, email, password: hashPassword(password, bcrypt) } })
       isCreated ? res.json({ status: 200, message: 'usuario registrado correctamente', user }) : res.json({ status: 404, message: 'esta cuenta ya existe' })
     }
   } catch (err) {
